requirements.txt
APScheduler==3.7.0
Pillow==9.5.0
aiohttp
aiofiles
asyncio
bs4
emoji==1.6.0
gitpython
google
gpytranslate
gtts
lottie
psycopg2-binary
pymediainfo
pyrogram
pytgcalls==3.0.0.dev22
python-dotenv
requests
sqlalchemy==1.3.23
telegraph
tgcrypto
tracemoepy
opencv-python-headless
urllib3
uvloop
